requirements.txt
babel==2.16.0
bracex==2.5
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
defusedxml==0.7.1
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-awesome-pages-plugin==2.9.3
mkdocs-get-deps==0.2.0
mkdocs-git-committers-plugin-2==2.3.0
mkdocs-material==9.5.34
mkdocs-material-extensions==1.3.1
natsort==8.4.0
packaging==24.1
paginate==0.5.7
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
Pygments==2.18.0
pymdown-extensions==10.9
python-dateutil==2.9.0.post0
python-frontmatter==1.1.0
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.7.24
requests==2.32.3
setuptools==74.1.0
six==1.16.0
smmap==5.0.1
tinycss2==1.3.0
urllib3==2.2.2
watchdog==5.0.1
wcmatch==9.0
webencodings==0.5.1
wheel==0.44.0
yamllint==1.35.1
